Don't use the numerical argument (silently) anymore, to switch to the modules db, but use an explicit with-modules-database.
That means, you cannot just the full commands and temporarily to the
modules database switch with C-u anymore.
Instead, you have to switch manually with xdmp-select-modules-database
and back to the previous database (i.e., with xdmp-select-database),
or, in your code, programatically with (with-modules-database ...).

Using the selector still works, of course. So probably just use that. :)